MAR-213 - some refactor
